mycelix-workspace/mycelix-core/0TML/pyproject.toml
[tool.poetry]
name = "zerotrustml"
version = "1.0.0"
description = "Zero-TrustML Holochain - Decentralized Federated Learning with Economic Incentives"
authors = ["Luminous Dynamics <dev@luminousdynamics.org>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/luminous-dynamics/0TML"
repository = "https://github.com/luminous-dynamics/0TML"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
numpy = ">=2.4.2"
torch = ">=2.10.0"
websockets = ">=11.0"
aiohttp = ">=3.8.0"
cryptography = ">=41.0.0"
prometheus-client = { version = ">=0.17.0", optional = true }
grafana-api = { version = ">=1.0.3", optional = true }
torchvision = "^0.25.0"
scipy = "^1.16.2"
matplotlib = "^3.10.7"
pyjwt = "^2.10.1"
zstandard = "^0.25.0"
lz4 = "^4.4.4"
redis = "^6.4.0"
asyncpg = "^0.30.0"
scikit-learn = "^1.8.0"
pyyaml = "^6.0"
base58 = "^2.1.1"
requests = "^2.31.0"
msgpack = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
black = ">=23.0.0"
mypy = ">=1.5.0"
ruff = ">=0.0.290"
web3 = "^7.14.0"

[tool.poetry.extras]
monitoring = ["prometheus-client", "grafana-api"]

[tool.poetry.scripts]
zerotrustml = "zerotrustml.cli:main"
zerotrustml-node = "zerotrustml.node_cli:main"
zerotrustml-monitor = "zerotrustml.monitoring.cli:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
